requirements.txt
streamlit
python-dotenv
pymupdf
pillow
pytesseract
gtts
